cert-rotation: apply pod workarounds only for 4.15 and lower

+++ ci-operator/step-registry/openshift/e2e/cert-rotation/prepare-for-shutdown/openshift-e2e-cert-rotation-prepare-for-shutdown-commands.sh
@@ -138,15 +138,19 @@ function wait-for-valid-lb-ext-kubeconfig {
 
 function pod-restart-workarounds {
   export KUBECONFIG=/tmp/lb-ext.kubeconfig
-  # Workaround for https://issues.redhat.com/browse/OCPBUGS-28735
-  # Restart OVN / Multus before proceeding
-  oc --request-timeout=5s -n openshift-multus delete pod -l app=multus --force --grace-period=0
-  oc --request-timeout=5s -n openshift-ovn-kubernetes delete pod -l app=ovnkube-node --force --grace-period=0
-  oc --request-timeout=5s -n openshift-ovn-kubernetes delete pod -l app=ovnkube-control-plane --force --grace-period=0
-  # Workaround for https://issues.redhat.com/browse/OCPBUGS-15827
-  # Restart console and console-operator pods
-  oc --request-timeout=5s -n openshift-console-operator delete pod --all --force --grace-period=0
-  oc --request-timeout=5s -n openshift-console delete pod --all --force --grace-period=0
+  ocp_minor_version=$(oc version -o json | jq -r '.openshiftVersion' | cut -d '.' -f2)
+
+  if ocp_minor_version <=15; then
+    # Workaround for https://issues.redhat.com/browse/OCPBUGS-28735
+    # Restart OVN / Multus before proceeding
+    oc --request-timeout=5s -n openshift-multus delete pod -l app=multus --force --grace-period=0
+    oc --request-timeout=5s -n openshift-ovn-kubernetes delete pod -l app=ovnkube-node --force --grace-period=0
+    oc --request-timeout=5s -n openshift-ovn-kubernetes delete pod -l app=ovnkube-control-plane --force --grace-period=0
+    # Workaround for https://issues.redhat.com/browse/OCPBUGS-15827
+    # Restart console and console-operator pods
+    oc --request-timeout=5s -n openshift-console-operator delete pod --all --force --grace-period=0
+    oc --request-timeout=5s -n openshift-console delete pod --all --force --grace-period=0
+  fi
 }
 
 function prepull-tools-image-for-gather-step {
